@@ -26,23 +26,24 @@
             "email": "d.bannert@anolilab.de",
             "homepage": "http://www.anolilab.de",
             "role": "Developer"
+        },
+        {
+            "name": "Narrowspark Community",
+            "homepage": "https://narrowspark.com/contributors"
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.3",
         "ext-curl": "*",
         "ext-fileinfo": "*",
-        "league/flysystem": "^1.0.44",
-        "spatie/macroable": "^1.0",
-        "symfony/filesystem": "^4.2.0",
-        "symfony/finder": "^4.2.0",
-        "symfony/polyfill-mbstring": "^1.8.0",
-        "viserio/contract": "self.version",
+        "league/flysystem": "^1.0.55",
+        "symfony/filesystem": "^4.3.5 || ^5.0.0",
+        "symfony/finder": "^4.3.5 || ^5.0.0",
+        "viserio/filesystem-contract": "self.version",
         "viserio/manager": "self.version",
         "viserio/options-resolver": "self.version"
     },
     "require-dev": {
-        "container-interop/service-provider": "^0.4",
         "league/flysystem-aws-s3-v3": "^1.0.19",
         "league/flysystem-cached-adapter": "^1.0.6",
         "league/flysystem-dropbox": "^1.0",
@@ -50,16 +51,19 @@
         "league/flysystem-vfs": "^1.0",
         "league/flysystem-webdav": "^1.0",
         "league/flysystem-ziparchive": "^1.0",
-        "mikey179/vfsStream": "^1.6",
-        "mockery/mockery": "^1.2.0",
-        "narrowspark/testing-helper": "^7.0.0",
+        "mikey179/vfsstream": "^1.6.7",
+        "mockery/mockery": "^1.2.3",
+        "narrowspark/coding-standard": "^3.1.0",
+        "narrowspark/testing-helper": "^8.0.1",
         "paragonie/halite": "^4.4.2",
-        "phpunit/phpunit": "^7.5.0",
+        "phpunit/phpunit": "8.2.*",
+        "slam/phpstan-extensions": "^3.6.1",
+        "spatie/flysystem-dropbox": "^1.1.0",
         "viserio/cache": "self.version",
-        "viserio/container": "self.version"
+        "viserio/container": "self.version",
+        "viserio/phpstan-bridge": "self.version"
     },
     "suggest": {
-        "container-interop/service-provider": "Required to use service-provider (^0.4).",
         "league/flysystem-aws-s3-v3": "Required to use the Flysystem S3 adapter (^1.0.19).",
         "league/flysystem-cached-adapter": "Required to use the Flysystem Caching adapter (^1.0.6).",
         "league/flysystem-dropbox": "Required to use the Flysystem Dropbox adapter (^1.0).",
@@ -69,7 +73,7 @@
         "league/flysystem-webdav": "Required to use the Flysystem WebDav adapter (^1.0).",
         "league/flysystem-ziparchive": "Required to use the Flysystem ZipArchive adapter (^1.0).",
         "paragonie/halite": "Required to use the encryption for files (^4.4.2).",
-        "viserio/statical-proxy": "Required to use the proxy instances (self.version)."
+        "viserio/container": "Required to use service-provider (self.version)."
     },
     "extra": {
         "branch-alias": {
