@@ -1,22 +1,32 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Adapter;
 
 use Aws\S3\S3Client;
 use League\Flysystem\AdapterInterface;
 use League\Flysystem\AwsS3v3\AwsS3Adapter as AwsS3v3;
-use Viserio\Component\Contract\Filesystem\Connector as ConnectorContract;
-use Viserio\Component\Contract\Filesystem\Exception\InvalidArgumentException;
-use Viserio\Component\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 use Viserio\Component\OptionsResolver\Traits\OptionsResolverTrait;
+use Viserio\Contract\Filesystem\Connector as ConnectorContract;
+use Viserio\Contract\Filesystem\Exception\InvalidArgumentException;
+use Viserio\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
+use Viserio\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 
-final class AwsS3Connector implements
-    ConnectorContract,
-    RequiresConfigContract,
+final class AwsS3Connector implements ConnectorContract,
     ProvidesDefaultOptionsContract,
+    RequiresConfigContract,
     RequiresMandatoryOptionsContract,
     RequiresValidatedConfigContract
 {
@@ -59,7 +69,7 @@ final class AwsS3Connector implements
     public static function getDefaultOptions(): array
     {
         return [
-            'prefix'  => null,
+            'prefix' => null,
             'options' => [],
         ];
     }
@@ -70,10 +80,10 @@ final class AwsS3Connector implements
     public static function getOptionValidators(): array
     {
         return [
-            'prefix'  => ['string', 'null'],
+            'prefix' => ['string', 'null'],
             'options' => ['array'],
-            'bucket'  => ['string'],
-            'auth'    => [
+            'bucket' => ['string'],
+            'auth' => [
                 'credentials' => static function ($value): void {
                     if (! \is_array($value)) {
                         throw new InvalidArgumentException('.');
