@@ -1,17 +1,27 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Adapter;
 
 use League\Flysystem\Adapter\Ftp;
 use League\Flysystem\AdapterInterface;
-use Viserio\Component\Contract\Filesystem\Connector as ConnectorContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 use Viserio\Component\OptionsResolver\Traits\OptionsResolverTrait;
+use Viserio\Contract\Filesystem\Connector as ConnectorContract;
+use Viserio\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
+use Viserio\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 
-final class FtpConnector implements
-    ConnectorContract,
+final class FtpConnector implements ConnectorContract,
     RequiresConfigContract,
     RequiresMandatoryOptionsContract,
     RequiresValidatedConfigContract
@@ -54,8 +64,8 @@ final class FtpConnector implements
     public static function getOptionValidators(): array
     {
         return [
-            'host'     => ['string'],
-            'port'     => ['string', 'int'],
+            'host' => ['string'],
+            'port' => ['string', 'int'],
             'username' => ['string'],
             'password' => ['string'],
         ];
