@@ -1,20 +1,30 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Adapter;
 
 use League\Flysystem\Adapter\Local;
 use League\Flysystem\AdapterInterface;
-use Viserio\Component\Contract\Filesystem\Connector as ConnectorContract;
-use Viserio\Component\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 use Viserio\Component\OptionsResolver\Traits\OptionsResolverTrait;
+use Viserio\Contract\Filesystem\Connector as ConnectorContract;
+use Viserio\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
+use Viserio\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 
-final class LocalConnector implements
-    ConnectorContract,
-    RequiresConfigContract,
+final class LocalConnector implements ConnectorContract,
     ProvidesDefaultOptionsContract,
+    RequiresConfigContract,
     RequiresMandatoryOptionsContract,
     RequiresValidatedConfigContract
 {
@@ -53,9 +63,9 @@ final class LocalConnector implements
     public static function getDefaultOptions(): array
     {
         return [
-            'write_flags'   => \LOCK_EX,
+            'write_flags' => \LOCK_EX,
             'link_handling' => Local::DISALLOW_LINKS,
-            'permissions'   => [],
+            'permissions' => [],
         ];
     }
 
@@ -65,17 +75,17 @@ final class LocalConnector implements
     public static function getOptionValidators(): array
     {
         return [
-            'host'     => ['string'],
-            'port'     => ['string', 'int'],
-            'username' => ['string'],
-            'password' => ['string'],
+            'path' => ['string'],
+            'write_flags' => ['int'],
+            'link_handling' => ['int'],
+            'permissions' => ['array'],
         ];
     }
 
     /**
      * Establish a connection.
      *
-     * @throws \Viserio\Component\Contract\OptionsResolver\Exception\InvalidArgumentException On wrong configuration
+     * @throws \Viserio\Contract\OptionsResolver\Exception\InvalidArgumentException On wrong configuration
      *
      * @return \League\Flysystem\Adapter\Local
      */
