@@ -1,21 +1,31 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Adapter;
 
 use League\Flysystem\AdapterInterface;
 use League\Flysystem\WebDAV\WebDAVAdapter;
 use Sabre\DAV\Client;
-use Viserio\Component\Contract\Filesystem\Connector as ConnectorContract;
-use Viserio\Component\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
-use Viserio\Component\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 use Viserio\Component\OptionsResolver\Traits\OptionsResolverTrait;
+use Viserio\Contract\Filesystem\Connector as ConnectorContract;
+use Viserio\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresConfig as RequiresConfigContract;
+use Viserio\Contract\OptionsResolver\RequiresMandatoryOptions as RequiresMandatoryOptionsContract;
+use Viserio\Contract\OptionsResolver\RequiresValidatedConfig as RequiresValidatedConfigContract;
 
-final class WebDavConnector implements
-    ConnectorContract,
-    RequiresConfigContract,
+final class WebDavConnector implements ConnectorContract,
     ProvidesDefaultOptionsContract,
+    RequiresConfigContract,
     RequiresMandatoryOptionsContract,
     RequiresValidatedConfigContract
 {
@@ -56,7 +66,7 @@ final class WebDavConnector implements
     public static function getDefaultOptions(): array
     {
         return [
-            'prefix'            => null,
+            'prefix' => null,
             'use_streamed_copy' => true,
         ];
     }
@@ -70,8 +80,8 @@ final class WebDavConnector implements
             'auth' => [
                 'baseUri' => ['string'],
             ],
-            'prefix'              => ['string', 'null'],
-            'use_streamed_copy'   => ['bool'],
+            'prefix' => ['string', 'null'],
+            'use_streamed_copy' => ['bool'],
         ];
     }
 
