@@ -1,5 +1,16 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Tests\Adapter;
 
 use PHPUnit\Framework\TestCase;
@@ -8,6 +19,8 @@ use Viserio\Component\Filesystem\Adapter\DropboxConnector;
 
 /**
  * @internal
+ *
+ * @small
  */
 final class DropboxConnectorTest extends TestCase
 {
@@ -19,18 +32,18 @@ final class DropboxConnectorTest extends TestCase
 
         $return = $connector->connect();
 
-        $this->assertInstanceOf(DropboxAdapter::class, $return);
+        self::assertInstanceOf(DropboxAdapter::class, $return);
     }
 
     public function testConnectWithPrefix(): void
     {
         $connector = new DropboxConnector([
-            'token'  => 'your-token',
+            'token' => 'your-token',
             'prefix' => 'your-prefix',
         ]);
 
         $return = $connector->connect();
 
-        $this->assertInstanceOf(DropboxAdapter::class, $return);
+        self::assertInstanceOf(DropboxAdapter::class, $return);
     }
 }
