@@ -1,5 +1,16 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Tests\Adapter;
 
 use League\Flysystem\WebDAV\WebDAVAdapter;
@@ -8,6 +19,8 @@ use Viserio\Component\Filesystem\Adapter\WebDavConnector;
 
 /**
  * @internal
+ *
+ * @small
  */
 final class WebDavConnectorTest extends TestCase
 {
@@ -15,7 +28,7 @@ final class WebDavConnectorTest extends TestCase
     {
         $connector = new WebDavConnector([
             'auth' => [
-                'baseUri'  => 'http://example.org/dav/',
+                'baseUri' => 'http://example.org/dav/',
             ],
             'userName' => 'your-username',
             'password' => 'your-password',
@@ -23,22 +36,22 @@ final class WebDavConnectorTest extends TestCase
 
         $return = $connector->connect();
 
-        $this->assertInstanceOf(WebDAVAdapter::class, $return);
+        self::assertInstanceOf(WebDAVAdapter::class, $return);
     }
 
     public function testConnectWithPrefix(): void
     {
         $connector = new WebDavConnector([
             'auth' => [
-                'baseUri'  => 'http://example.org/dav/',
+                'baseUri' => 'http://example.org/dav/',
             ],
             'userName' => 'your-username',
             'password' => 'your-password',
-            'prefix'   => 'your-prefix',
+            'prefix' => 'your-prefix',
         ]);
 
         $return = $connector->connect();
 
-        $this->assertInstanceOf(WebDAVAdapter::class, $return);
+        self::assertInstanceOf(WebDAVAdapter::class, $return);
     }
 }
