@@ -1,27 +1,36 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Tests\Encryption;
 
 use ParagonIE\Halite\KeyFactory;
 use PHPUnit\Framework\TestCase;
-use Viserio\Component\Contract\Filesystem\Exception\FileNotFoundException;
 use Viserio\Component\Filesystem\Adapter\LocalConnector;
 use Viserio\Component\Filesystem\Encryption\EncryptionWrapper;
 use Viserio\Component\Filesystem\FilesystemAdapter;
+use Viserio\Contract\Filesystem\Exception\FileNotFoundException;
 
 /**
  * @internal
+ *
+ * @small
  */
 final class EncryptionWrapperTest extends TestCase
 {
-    /**
-     * @var string
-     */
+    /** @var string */
     private $root;
 
-    /**
-     * @var \Viserio\Component\Filesystem\Encryption\EncryptionWrapper
-     */
+    /** @var \Viserio\Component\Filesystem\Encryption\EncryptionWrapper */
     private $adapter;
 
     /**
@@ -29,12 +38,12 @@ final class EncryptionWrapperTest extends TestCase
      */
     protected function setUp(): void
     {
-        if (\mb_stripos(\PHP_OS, 'win') === 0) {
-            $this->markTestSkipped('@Todo fix this test on windows.');
+        if (\stripos(\PHP_OS, 'win') === 0) {
+            self::markTestSkipped('@Todo fix this test on windows.');
         }
 
         $this->root = __DIR__ . \DIRECTORY_SEPARATOR . 'stubs';
-        $connector  = new LocalConnector(['path' => $this->root]);
+        $connector = new LocalConnector(['path' => $this->root]);
 
         $adapter = $connector->connect();
 
@@ -68,28 +77,28 @@ final class EncryptionWrapperTest extends TestCase
 
         $temp = \fopen($filePath, 'rb');
 
-        $this->assertTrue($this->adapter->writeStream('encrypt.txt', $temp));
-        $this->assertSame('dummy', \stream_get_contents($this->adapter->readStream('encrypt.txt')));
+        self::assertTrue($this->adapter->writeStream('encrypt.txt', $temp));
+        self::assertSame('dummy', \stream_get_contents($this->adapter->readStream('encrypt.txt')));
     }
 
     public function testWrite(): void
     {
-        $this->assertTrue($this->adapter->write('encrypt.txt', 'dummy'));
-        $this->assertSame('dummy', $this->adapter->read('encrypt.txt'));
+        self::assertTrue($this->adapter->write('encrypt.txt', 'dummy'));
+        self::assertSame('dummy', $this->adapter->read('encrypt.txt'));
     }
 
     public function testUpdate(): void
     {
-        $this->assertTrue($this->adapter->write('encrypt_update.txt', 'dummy'));
-        $this->assertTrue($this->adapter->update('encrypt_update.txt', 'file'));
+        self::assertTrue($this->adapter->write('encrypt_update.txt', 'dummy'));
+        self::assertTrue($this->adapter->update('encrypt_update.txt', 'file'));
 
-        $this->assertSame('file', $this->adapter->read('encrypt_update.txt'));
+        self::assertSame('file', $this->adapter->read('encrypt_update.txt'));
     }
 
     public function testPut(): void
     {
-        $this->assertTrue($this->adapter->put('encrypt_put.txt', 'file'));
-        $this->assertSame('file', $this->adapter->read('encrypt_put.txt'));
+        self::assertTrue($this->adapter->put('encrypt_put.txt', 'file'));
+        self::assertSame('file', $this->adapter->read('encrypt_put.txt'));
 
         $filePath = $this->root . \DIRECTORY_SEPARATOR . 'dummy.text';
 
@@ -97,11 +106,11 @@ final class EncryptionWrapperTest extends TestCase
 
         $temp = \fopen($filePath, 'rb');
 
-        $this->assertTrue($this->adapter->put('encrypt_put.txt', $temp));
-        $this->assertSame('dummy', $this->adapter->read('encrypt_put.txt'));
+        self::assertTrue($this->adapter->put('encrypt_put.txt', $temp));
+        self::assertSame('dummy', $this->adapter->read('encrypt_put.txt'));
 
-        $this->assertTrue($this->adapter->put('encrypt_put2.txt', $temp));
-        $this->assertSame('dummy', $this->adapter->read('encrypt_put.txt'));
+        self::assertTrue($this->adapter->put('encrypt_put2.txt', $temp));
+        self::assertSame('dummy', $this->adapter->read('encrypt_put.txt'));
     }
 
     public function testUpdateStream(): void
@@ -112,8 +121,8 @@ final class EncryptionWrapperTest extends TestCase
 
         $temp = \fopen($filePath, 'rb');
 
-        $this->assertTrue($this->adapter->updateStream('encrypt_u_stream.txt', $temp));
-        $this->assertSame('dummy', $this->adapter->read('encrypt_u_stream.txt'));
+        self::assertTrue($this->adapter->updateStream('encrypt_u_stream.txt', $temp));
+        self::assertSame('dummy', $this->adapter->read('encrypt_u_stream.txt'));
 
         $filePath = $this->root . \DIRECTORY_SEPARATOR . 'dummy.text';
 
@@ -121,8 +130,8 @@ final class EncryptionWrapperTest extends TestCase
 
         $temp = \fopen($filePath, 'rb');
 
-        $this->assertTrue($this->adapter->updateStream('encrypt_u_stream.txt', $temp));
-        $this->assertSame('file', $this->adapter->read('encrypt_u_stream.txt'));
+        self::assertTrue($this->adapter->updateStream('encrypt_u_stream.txt', $temp));
+        self::assertSame('file', $this->adapter->read('encrypt_u_stream.txt'));
     }
 
     public function testRead(): void
