@@ -1,5 +1,16 @@
 <?php
+
 declare(strict_types=1);
+
+/**
+ * This file is part of Narrowspark Framework.
+ *
+ * (c) Daniel Bannert <d.bannert@anolilab.de>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Viserio\Component\Filesystem\Tests;
 
 use League\Flysystem\Util;
@@ -8,23 +19,21 @@ use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Finder\SplFileInfo;
-use Viserio\Component\Contract\Filesystem\Exception\FileNotFoundException;
-use Viserio\Component\Contract\Filesystem\Exception\IOException;
 use Viserio\Component\Filesystem\Filesystem;
+use Viserio\Contract\Filesystem\Exception\FileNotFoundException;
+use Viserio\Contract\Filesystem\Exception\IOException;
 
 /**
  * @internal
+ *
+ * @small
  */
 final class FilesystemTest extends TestCase
 {
-    /**
-     * @var \org\bovigo\vfs\vfsStreamDirectory
-     */
+    /** @var \org\bovigo\vfs\vfsStreamDirectory */
     private $root;
 
-    /**
-     * @var \Viserio\Component\Filesystem\Filesystem
-     */
+    /** @var \Viserio\Component\Filesystem\Filesystem */
     private $files;
 
     /**
@@ -32,7 +41,7 @@ final class FilesystemTest extends TestCase
      */
     protected function setUp(): void
     {
-        $this->root  = vfsStream::setup();
+        $this->root = vfsStream::setup();
         $this->files = new Filesystem();
     }
 
@@ -40,7 +49,7 @@ final class FilesystemTest extends TestCase
     {
         $file = vfsStream::newFile('temp.txt')->withContent('Foo Bar')->at($this->root);
 
-        $this->assertEquals('Foo Bar', $this->files->read($file->url()));
+        self::assertEquals('Foo Bar', $this->files->read($file->url()));
     }
 
     public function testReadStreamToThrowException(): void
@@ -64,7 +73,7 @@ final class FilesystemTest extends TestCase
         $this->files->update($file->url(), 'Hello World');
         $this->files->update($file->url(), 'Hello World2');
 
-        $this->assertStringEqualsFile($file->url(), 'Hello World2');
+        self::assertStringEqualsFile($file->url(), 'Hello World2');
     }
 
     public function testPutStoresFiles(): void
@@ -73,7 +82,7 @@ final class FilesystemTest extends TestCase
 
         $this->files->put($file->url(), 'Hello World');
 
-        $this->assertStringEqualsFile($file->url(), 'Hello World');
+        self::assertStringEqualsFile($file->url(), 'Hello World');
     }
 
     public function testUpdateToThrowException(): void
@@ -87,41 +96,41 @@ final class FilesystemTest extends TestCase
     {
         $this->root->addChild(new vfsStreamDirectory('temp'));
 
-        $dir  = $this->root->getChild('temp');
+        $dir = $this->root->getChild('temp');
         $file = vfsStream::newFile('bar.txt')->withContent('bar')->at($dir);
 
-        $this->assertDirectoryExists($dir->url());
-        $this->assertFalse($this->files->deleteDirectory($file->url()));
+        self::assertDirectoryExists($dir->url());
+        self::assertFalse($this->files->deleteDirectory($file->url()));
 
         $this->files->deleteDirectory($dir->url());
 
-        $this->assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'temp'));
-        $this->assertFileNotExists($file->url());
+        self::assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'temp'));
+        self::assertFileNotExists($file->url());
     }
 
     public function testCleanDirectory(): void
     {
         $this->root->addChild(new vfsStreamDirectory('tempdir'));
 
-        $dir  = $this->root->getChild('tempdir');
+        $dir = $this->root->getChild('tempdir');
         $file = vfsStream::newFile('tempfoo.txt')->withContent('tempfoo')->at($dir);
 
-        $this->assertFalse($this->files->cleanDirectory($file->url()));
+        self::assertFalse($this->files->cleanDirectory($file->url()));
         $this->files->cleanDirectory($dir->url());
 
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tempdir'));
-        $this->assertFileNotExists($file->url());
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tempdir'));
+        self::assertFileNotExists($file->url());
     }
 
     public function testDeleteRemovesFiles(): void
     {
         $file = vfsStream::newFile('unlucky.txt')->withContent('So sad')->at($this->root);
 
-        $this->assertTrue($this->files->has($file->url()));
+        self::assertTrue($this->files->has($file->url()));
 
         $this->files->delete([$file->url()]);
 
-        $this->assertFalse($this->files->has($file->url()));
+        self::assertFalse($this->files->has($file->url()));
     }
 
     public function testMoveMovesFiles(): void
@@ -131,59 +140,59 @@ final class FilesystemTest extends TestCase
 
         $this->files->move($file->url(), $rock);
 
-        $this->assertFileExists($rock);
-        $this->assertStringEqualsFile($rock, 'pop');
-        $this->assertFileNotExists($this->root->url() . \DIRECTORY_SEPARATOR . 'pop.txt');
+        self::assertFileExists($rock);
+        self::assertStringEqualsFile($rock, 'pop');
+        self::assertFileNotExists($this->root->url() . \DIRECTORY_SEPARATOR . 'pop.txt');
     }
 
     public function testGetExtensionReturnsExtension(): void
     {
         $file = vfsStream::newFile('rock.csv')->withContent('pop,rock')->at($this->root);
 
-        $this->assertEquals('csv', $this->files->getExtension($file->url()));
+        self::assertEquals('csv', $this->files->getExtension($file->url()));
     }
 
     public function testGetMimeTypeOutputsMimeType(): void
     {
         if (! \class_exists('Finfo')) {
-            $this->markTestSkipped('The PHP extension fileinfo is not installed.');
+            self::markTestSkipped('The PHP extension fileinfo is not installed.');
         }
 
         $file = vfsStream::newFile('foo.txt')->withContent('foo')->at($this->root);
 
-        $this->assertEquals('text/plain', $this->files->getMimetype($file->url()));
+        self::assertEquals('text/plain', $this->files->getMimetype($file->url()));
     }
 
     public function testGetSizeOutputsSize(): void
     {
         $content = LargeFileContent::withKilobytes(2);
-        $file    = vfsStream::newFile('2kb.txt')->withContent($content)->at($this->root);
+        $file = vfsStream::newFile('2kb.txt')->withContent($content)->at($this->root);
 
-        $this->assertEquals($file->size(), $this->files->getSize($file->url()));
+        self::assertEquals($file->size(), $this->files->getSize($file->url()));
     }
 
     public function testIsDirectory(): void
     {
         $this->root->addChild(new vfsStreamDirectory('assets'));
-        $dir  = $this->root->getChild('assets');
+        $dir = $this->root->getChild('assets');
         $file = vfsStream::newFile('foo.txt')->withContent('foo')->at($this->root);
 
-        $this->assertTrue($this->files->isDirectory($dir->url()));
-        $this->assertFalse($this->files->isDirectory($file->url()));
+        self::assertTrue($this->files->isDirectory($dir->url()));
+        self::assertFalse($this->files->isDirectory($file->url()));
     }
 
     public function testAllFilesFindsFiles(): void
     {
         $this->root->addChild(new vfsStreamDirectory('languages'));
 
-        $dir   = $this->root->getChild('languages');
+        $dir = $this->root->getChild('languages');
         $file1 = vfsStream::newFile('php.txt')->withContent('PHP')->at($dir);
         $file2 = vfsStream::newFile('c.txt')->withContent('C')->at($dir);
 
         $allFiles = $this->files->allFiles($dir->url());
 
-        $this->assertContains($file1->getName(), $allFiles[0]);
-        $this->assertContains($file2->getName(), $allFiles[1]);
+        self::assertStringContainsString($file1->getName(), $allFiles[0]);
+        self::assertStringContainsString($file2->getName(), $allFiles[1]);
     }
 
     public function testDirectoriesFindsDirectories(): void
@@ -193,20 +202,20 @@ final class FilesystemTest extends TestCase
 
         $directories = $this->files->directories($this->root->url());
 
-        $this->assertContains('vfs://root' . \DIRECTORY_SEPARATOR . 'languages', $directories[0]);
-        $this->assertContains('vfs://root' . \DIRECTORY_SEPARATOR . 'music', $directories[1]);
+        self::assertStringContainsString('vfs://root' . \DIRECTORY_SEPARATOR . 'languages', $directories[0]);
+        self::assertStringContainsString('vfs://root' . \DIRECTORY_SEPARATOR . 'music', $directories[1]);
     }
 
     public function testCreateDirectory(): void
     {
         $this->files->createDirectory($this->root->url() . \DIRECTORY_SEPARATOR . 'test');
 
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'test'));
-        $this->assertEquals(0755, $this->root->getChild('test')->getPermissions());
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'test'));
+        self::assertEquals(0755, $this->root->getChild('test')->getPermissions());
 
         $this->files->createDirectory($this->root->url() . \DIRECTORY_SEPARATOR . 'test2', ['visibility' => 'private']);
 
-        $this->assertEquals(0700, $this->root->getChild('test2')->getPermissions());
+        self::assertEquals(0700, $this->root->getChild('test2')->getPermissions());
     }
 
     public function testCopy(): void
@@ -223,7 +232,7 @@ final class FilesystemTest extends TestCase
             $this->root->getChild('copy2')->url() . \DIRECTORY_SEPARATOR . 'copy.txt'
         );
 
-        $this->assertSame(
+        self::assertSame(
             'copy1',
             $this->files->read(
                 $this->root->getChild('copy2')->url() . \DIRECTORY_SEPARATOR . 'copy.txt'
@@ -269,18 +278,18 @@ final class FilesystemTest extends TestCase
             ->withContent('copy')
             ->at($dir);
 
-        $this->assertSame('public', $this->files->getVisibility($dir->url()));
-        $this->assertSame('public', $this->files->getVisibility($file->url()));
+        self::assertSame('0777', $this->files->getVisibility($dir->url()));
+        self::assertSame('0666', $this->files->getVisibility($file->url()));
 
         $this->files->setVisibility($file->url(), 'private');
         $this->files->setVisibility($dir->url(), 'private');
 
-        $this->assertSame('private', $this->files->getVisibility($dir->url()));
-        $this->assertSame('private', $this->files->getVisibility($file->url()));
+        self::assertSame('private', $this->files->getVisibility($dir->url()));
+        self::assertSame('private', $this->files->getVisibility($file->url()));
 
         $this->files->setVisibility($file->url(), 'public');
 
-        $this->assertSame('public', $this->files->getVisibility($file->url()));
+        self::assertSame('public', $this->files->getVisibility($file->url()));
     }
 
     public function testSetVisibilityToThrowInvalidArgumentException(): void
@@ -306,12 +315,12 @@ final class FilesystemTest extends TestCase
 
         $this->files->write($file->url(), 'copy new');
 
-        $this->assertSame('copy new', $this->files->read($file->url()));
+        self::assertSame('copy new', $this->files->read($file->url()));
 
         $this->files->write($file->url(), 'copy new visibility', ['visibility' => 'private']);
 
-        $this->assertSame('copy new visibility', $this->files->read($file->url()));
-        $this->assertSame('private', $this->files->getVisibility($file->url()));
+        self::assertSame('copy new visibility', $this->files->read($file->url()));
+        self::assertSame('private', $this->files->getVisibility($file->url()));
     }
 
     public function testWriteStreamAndReadStream(): void
@@ -330,13 +339,13 @@ final class FilesystemTest extends TestCase
         $stream = $this->files->readStream($file->url());
 
         $contents = \stream_get_contents($stream);
-        $size     = Util::getStreamSize($stream);
+        $size = Util::getStreamSize($stream);
 
         \fclose($stream);
 
-        $this->assertSame(5, $size);
-        $this->assertSame('dummy', $contents);
-        $this->assertIsResource($this->files->readStream($file->url()));
+        self::assertSame(5, $size);
+        self::assertSame('dummy', $contents);
+        self::assertIsResource($this->files->readStream($file->url()));
     }
 
     public function testUpdateStream(): void
@@ -347,14 +356,14 @@ final class FilesystemTest extends TestCase
             ->withContent('copy')
             ->at($this->root->getChild('copy'));
 
-        $this->assertSame('copy', $this->files->read($file->url()));
+        self::assertSame('copy', $this->files->read($file->url()));
 
         $temp = \tmpfile();
 
         \fwrite($temp, 'dummy');
         \rewind($temp);
 
-        $this->assertTrue($this->files->updateStream(
+        self::assertTrue($this->files->updateStream(
             $file->url(),
             $temp,
             ['visibility' => 'public']
@@ -363,12 +372,12 @@ final class FilesystemTest extends TestCase
         $stream = $this->files->readStream($file->url());
 
         $contents = \stream_get_contents($stream);
-        $size     = Util::getStreamSize($stream);
+        $size = Util::getStreamSize($stream);
 
         \fclose($stream);
 
-        $this->assertSame(5, $size);
-        $this->assertSame('dummy', $contents);
+        self::assertSame(5, $size);
+        self::assertSame('dummy', $contents);
     }
 
     public function testGetMimetype(): void
@@ -381,7 +390,7 @@ final class FilesystemTest extends TestCase
             ->withContent('copy')
             ->at($dir);
 
-        $this->assertSame('text/plain', $this->files->getMimetype($file->url()));
+        self::assertSame('text/plain', $this->files->getMimetype($file->url()));
     }
 
     public function testGetMimetypeToThrowFileNotFoundException(): void
@@ -401,7 +410,7 @@ final class FilesystemTest extends TestCase
             ->withContent('copy')
             ->at($dir);
 
-        $this->assertSame(\date('F d Y H:i:s', \filemtime($file->url())), $this->files->getTimestamp($file->url()));
+        self::assertSame(\date('F d Y H:i:s', \filemtime($file->url())), $this->files->getTimestamp($file->url()));
     }
 
     public function testGetTimestampToThrowFileNotFoundException(): void
@@ -416,7 +425,7 @@ final class FilesystemTest extends TestCase
         $this->root->addChild(new vfsStreamDirectory('tmp'));
         $this->root->addChild(new vfsStreamDirectory('tmp2'));
 
-        $dir   = $this->root->getChild('tmp');
+        $dir = $this->root->getChild('tmp');
         $temp2 = $this->root->getChild('tmp2');
 
         vfsStream::newFile('foo.txt')
@@ -435,12 +444,12 @@ final class FilesystemTest extends TestCase
 
         $this->files->moveDirectory($dir->url(), $temp2->url());
 
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
-        $this->assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp'));
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
+        self::assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp'));
     }
 
     public function testMoveDirectoryMovesEntireDirectoryAndOverwrites(): void
@@ -448,7 +457,7 @@ final class FilesystemTest extends TestCase
         $this->root->addChild(new vfsStreamDirectory('tmp'));
         $this->root->addChild(new vfsStreamDirectory('tmp2'));
 
-        $dir   = $this->root->getChild('tmp');
+        $dir = $this->root->getChild('tmp');
         $temp2 = $this->root->getChild('tmp2');
 
         vfsStream::newFile('foo.txt')
@@ -474,19 +483,19 @@ final class FilesystemTest extends TestCase
 
         $this->files->moveDirectory($dir->url(), $temp2->url(), ['overwrite' => true]);
 
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
-        $this->assertFileNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo2.txt');
-        $this->assertFileNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar2.txt');
-        $this->assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp'));
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
+        self::assertFileNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo2.txt');
+        self::assertFileNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar2.txt');
+        self::assertDirectoryNotExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp'));
     }
 
     public function testCopyDirectoryReturnsFalseIfSourceIsntDirectory(): void
     {
-        $this->assertFalse($this->files->copyDirectory(\DIRECTORY_SEPARATOR . 'foo' . \DIRECTORY_SEPARATOR . 'bar' . \DIRECTORY_SEPARATOR . 'baz' . \DIRECTORY_SEPARATOR . 'breeze' . \DIRECTORY_SEPARATOR . 'boom', 'foo'));
+        self::assertFalse($this->files->copyDirectory(\DIRECTORY_SEPARATOR . 'foo' . \DIRECTORY_SEPARATOR . 'bar' . \DIRECTORY_SEPARATOR . 'baz' . \DIRECTORY_SEPARATOR . 'breeze' . \DIRECTORY_SEPARATOR . 'boom', 'foo'));
     }
 
     public function testCopyDirectoryMovesEntireDirectory(): void
@@ -494,7 +503,7 @@ final class FilesystemTest extends TestCase
         $this->root->addChild(new vfsStreamDirectory('tmp'));
         $this->root->addChild(new vfsStreamDirectory('tmp2'));
 
-        $dir   = $this->root->getChild('tmp');
+        $dir = $this->root->getChild('tmp');
         $temp2 = $this->root->getChild('tmp2');
 
         vfsStream::newFile('foo.txt')
@@ -513,11 +522,11 @@ final class FilesystemTest extends TestCase
 
         $this->files->copyDirectory($dir->url(), $temp2->url());
 
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
-        $this->assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
-        $this->assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2'));
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'foo.txt');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'bar.txt');
+        self::assertDirectoryExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested');
+        self::assertFileExists(vfsStream::url('root' . \DIRECTORY_SEPARATOR . 'tmp2') . \DIRECTORY_SEPARATOR . 'nested' . \DIRECTORY_SEPARATOR . 'baz.txt');
     }
 
     public function testFiles(): void
@@ -540,9 +549,9 @@ final class FilesystemTest extends TestCase
             ->withContent('baz')
             ->at($dir2);
 
-        $this->assertContains('bar.txt', $this->files->files($dir->url()));
-        $this->assertContains('foo.txt', $this->files->files($dir->url()));
-        $this->assertNotContains('foo2.txt', $this->files->files($dir->url()));
+        self::assertContains('bar.txt', $this->files->files($dir->url()));
+        self::assertContains('foo.txt', $this->files->files($dir->url()));
+        self::assertNotContains('foo2.txt', $this->files->files($dir->url()));
     }
 
     public function testAllDirectories(): void
@@ -552,7 +561,7 @@ final class FilesystemTest extends TestCase
 
         $arr = $this->files->allDirectories($this->root->url());
 
-        $this->assertInstanceOf(SplFileInfo::class, $arr[0]);
+        self::assertInstanceOf(SplFileInfo::class, $arr[0]);
     }
 
     public function testAppendOnExistingFile(): void
@@ -561,7 +570,7 @@ final class FilesystemTest extends TestCase
 
         $this->files->append($file->url(), ' test');
 
-        $this->assertEquals('Foo Bar test', $this->files->read($file->url()));
+        self::assertEquals('Foo Bar test', $this->files->read($file->url()));
     }
 
     public function testAppend(): void
@@ -570,7 +579,7 @@ final class FilesystemTest extends TestCase
 
         $this->files->append($url, 'test');
 
-        $this->assertEquals('test', $this->files->read($url));
+        self::assertEquals('test', $this->files->read($url));
     }
 
     public function testAppendStreamOnExistingFile(): void
@@ -586,18 +595,18 @@ final class FilesystemTest extends TestCase
         $stream = $this->files->readStream($file->url());
 
         $contents = \stream_get_contents($stream);
-        $size     = Util::getStreamSize($stream);
+        $size = Util::getStreamSize($stream);
 
         \fclose($stream);
 
-        $this->assertSame(13, $size);
-        $this->assertSame('Foo Bar dummy', $contents);
-        $this->assertIsResource($this->files->readStream($file->url()));
+        self::assertSame(13, $size);
+        self::assertSame('Foo Bar dummy', $contents);
+        self::assertIsResource($this->files->readStream($file->url()));
     }
 
     public function testAppendStream(): void
     {
-        $url  = $this->root->url() . \DIRECTORY_SEPARATOR . 'file.php';
+        $url = $this->root->url() . \DIRECTORY_SEPARATOR . 'file.php';
         $temp = \tmpfile();
 
         \fwrite($temp, 'dummy');
@@ -608,12 +617,12 @@ final class FilesystemTest extends TestCase
         $stream = $this->files->readStream($url);
 
         $contents = \stream_get_contents($stream);
-        $size     = Util::getStreamSize($stream);
+        $size = Util::getStreamSize($stream);
 
         \fclose($stream);
 
-        $this->assertSame(5, $size);
-        $this->assertSame('dummy', $contents);
-        $this->assertIsResource($this->files->readStream($url));
+        self::assertSame(5, $size);
+        self::assertSame('dummy', $contents);
+        self::assertIsResource($this->files->readStream($url));
     }
 }
